stm32cube: update stm32h7rs to cube version V1.3.0
Update Cube version for STM32H7RSxx series
on https://github.com/STMicroelectronics
from version v1.2.0
to version v1.3.0

Sync stm32_assert.h with other SoCs HAL by including header file
stm32h7rsxx_hal_conf.h instead of defining assert_failed() macro
that is already defined by that header file.

Add references in README file to Zepĥyr STM32 HAL pull requests for
the main local patches.

+++ stm32cube/stm32h7rsxx/drivers/include/stm32h7rsxx_ll_rcc.h
@@ -87,10 +87,10 @@ extern "C" {
 
 #define LL_CLKSOURCE_REG(__CLKSOURCE__)     (((__CLKSOURCE__) >> LL_RCC_REG_SHIFT   ) & 0xFFUL)
 
-#define LL_CLKSOURCE(__REG__, __MSK__, __POS__, __CLK__) ((uint32_t)((((__MSK__) >> (__POS__)) << LL_RCC_MASK_SHIFT) | \
-                                                                     (( __POS__              ) << LL_RCC_POS_SHIFT)  | \
-                                                                     (( __REG__              ) << LL_RCC_REG_SHIFT)  | \
-                                                                     (((__CLK__) >> (__POS__)) << LL_RCC_CONFIG_SHIFT)))
+#define LL_CLKSOURCE(__REG__, __MSK__, __POS__, __CLK__) ((uint32_t)((((uint32_t)(__MSK__) >> (__POS__)) << LL_RCC_MASK_SHIFT) | \
+                                                                     (((uint32_t)(__POS__)   ) << LL_RCC_POS_SHIFT)  | \
+                                                                     (((uint32_t)(__REG__)   ) << LL_RCC_REG_SHIFT)  | \
+                                                                     (((uint32_t)(__CLK__) >> (__POS__)) << LL_RCC_CONFIG_SHIFT)))
 
 /**
   * @}
